server/requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain-community==0.3.27
python-dotenv==1.1.1
uvicorn==0.35.0
chromadb==0.4.22
python-multipart==0.0.20
requests==2.32.5
openai==1.57.0
